# Perf: KZG with fixed-argument pairing in affine coordinates