--- a/ecc/bls12-377/internal/fptower/e12_pairing.go
+++ b/ecc/bls12-377/internal/fptower/e12_pairing.go
@@ -63,6 +63,27 @@ func (z *E12) MulBy034(c0, c3, c4 *E2) *E12 {
 	return z
 }
 
+// MulBy34 multiplication by sparse element (1,0,0,c3,c4,0)
+func (z *E12) MulBy34(c3, c4 *E2) *E12 {
+
+	var a, b, d E6
+
+	a.Set(&z.C0)
+
+	b.Set(&z.C1)
+	b.MulBy01(c3, c4)
+
+	var d0 E2
+	d0.SetOne().Add(&d0, c3)
+	d.Add(&z.C0, &z.C1)
+	d.MulBy01(&d0, c4)
+
+	z.C1.Add(&a, &b).Neg(&z.C1).Add(&z.C1, &d)
+	z.C0.MulByNonResidue(&b).Add(&z.C0, &a)
+
+	return z
+}
+
 // Mul034By034 multiplication of sparse element (c0,0,0,c3,c4,0) by sparse element (d0,0,0,d3,d4,0)
 func Mul034By034(d0, d3, d4, c0, c3, c4 *E2) [5]E2 {
 	var z00, tmp, x0, x3, x4, x04, x03, x34 E2
@@ -85,6 +106,26 @@ func Mul034By034(d0, d3, d4, c0, c3, c4 *E2) [5]E2 {
 		Sub(&x34, &x3).
 		Sub(&x34, &x4)
 
+	z00.MulByNonResidue(&x4).
+		Add(&z00, &x0)
+
+	return [5]E2{z00, x3, x34, x03, x04}
+}
+
+// Mul34By34 multiplication of sparse element (1,0,0,c3,c4,0) by sparse element (1,0,0,d3,d4,0)
+func Mul34By34(d3, d4, c3, c4 *E2) [5]E2 {
+	var z00, tmp, x0, x3, x4, x04, x03, x34 E2
+	x3.Mul(c3, d3)
+	x4.Mul(c4, d4)
+	x04.Add(c4, d4)
+	x03.Add(c3, d3)
+	tmp.Add(c3, c4)
+	x34.Add(d3, d4).
+		Mul(&x34, &tmp).
+		Sub(&x34, &x3).
+		Sub(&x34, &x4)
+
+	x0.SetOne()
 	z00.MulByNonResidue(&x4).
 		Add(&z00, &x0)