--- a/ecc/bls12-377/kzg/kzg.go
+++ b/ecc/bls12-377/kzg/kzg.go
@@ -49,7 +49,7 @@ type ProvingKey struct {
 
 // VerifyingKey used to verify opening proofs
 type VerifyingKey struct {
-	G2 [2]bls12377.G2Affine // [G₂, [α]G₂ ]
+	Lines [2][2][len(bls12377.LoopCounter)]bls12377.LineEvaluationAff // precomputed pairing lines corresponding to G₂, [α]G₂
 }
 
 // SRS must be computed through MPC and comprises the ProvingKey and the VerifyingKey
@@ -116,13 +116,17 @@ func NewSRS(size uint64, bAlpha *big.Int) (*SRS, error) {
 				srs.Pk.G1[i] = g[i%4]
 			}
 		})
-		srs.Vk.G2[0] = gen2Aff
-		srs.Vk.G2[1].ScalarMultiplication(&srs.Vk.G2[0], &bt)
+		var btG2 bls12377.G2Affine
+		btG2.ScalarMultiplication(&gen2Aff, &bt)
+		srs.Vk.Lines[0] = bls12377.PrecomputeLines(gen2Aff)
+		srs.Vk.Lines[1] = bls12377.PrecomputeLines(btG2)
 		return &srs, nil
 	}
 	srs.Pk.G1[0] = gen1Aff
-	srs.Vk.G2[0] = gen2Aff
-	srs.Vk.G2[1].ScalarMultiplication(&gen2Aff, bAlpha)
+	var bAlphaG2 bls12377.G2Affine
+	bAlphaG2.ScalarMultiplication(&gen2Aff, bAlpha)
+	srs.Vk.Lines[0] = bls12377.PrecomputeLines(gen2Aff)
+	srs.Vk.Lines[1] = bls12377.PrecomputeLines(bAlphaG2)
 
 	alphas := make([]fr.Element, size-1)
 	alphas[0] = alpha
@@ -234,9 +238,9 @@ func Verify(commitment *Digest, proof *OpeningProof, point fr.Element, vk Verify
 	totalG1Aff.FromJacobian(&totalG1)
 
 	// e([f(α)-f(a)+aH(α)]G₁], G₂).e([-H(α)]G₁, [α]G₂) == 1
-	check, err := bls12377.PairingCheck(
+	check, err := bls12377.PairingCheckFixedQ(
 		[]bls12377.G1Affine{totalG1Aff, negH},
-		[]bls12377.G2Affine{vk.G2[0], vk.G2[1]},
+		[][2][len(bls12377.LoopCounter)]bls12377.LineEvaluationAff{vk.Lines[0], vk.Lines[1]},
 	)
 	if err != nil {
 		return err
@@ -493,9 +497,9 @@ func BatchVerifyMultiPoints(digests []Digest, proofs []OpeningProof, points []fr
 
 	// pairing check
 	// e([∑ᵢλᵢ(fᵢ(α) - fᵢ(pᵢ) + pᵢHᵢ(α))]G₁, G₂).e([-∑ᵢλᵢ[Hᵢ(α)]G₁), [α]G₂)
-	check, err := bls12377.PairingCheck(
+	check, err := bls12377.PairingCheckFixedQ(
 		[]bls12377.G1Affine{foldedDigests, foldedQuotients},
-		[]bls12377.G2Affine{vk.G2[0], vk.G2[1]},
+		[][2][len(bls12377.LoopCounter)]bls12377.LineEvaluationAff{vk.Lines[0], vk.Lines[1]},
 	)
 	if err != nil {
 		return err