--- a/ecc/bls12-377/kzg/marshal.go
+++ b/ecc/bls12-377/kzg/marshal.go
@@ -57,7 +57,6 @@ func (vk *VerifyingKey) writeTo(w io.Writer, options ...func(*bls12377.Encoder))
 	toEncode := []interface{}{
 		&vk.G2[0],
 		&vk.G2[1],
-		&vk.G1,
 	}
 
 	for _, v := range toEncode {
@@ -110,7 +109,6 @@ func (vk *VerifyingKey) ReadFrom(r io.Reader) (int64, error) {
 	toDecode := []interface{}{
 		&vk.G2[0],
 		&vk.G2[1],
-		&vk.G1,
 	}
 
 	for _, v := range toDecode {