--- a/ecc/bls12-377/pairing.go
+++ b/ecc/bls12-377/pairing.go
@@ -417,7 +417,10 @@ func MillerLoopFixedQ(P []G1Affine, lines [][2][len(LoopCounter) - 1]LineEvaluat
 	yInv := make([]fp.Element, n)
 	xNegOverY := make([]fp.Element, n)
 	for k := 0; k < n; k++ {
-		yInv[k].Inverse(&P[k].Y)
+		yInv[k].Set(&P[k].Y)
+	}
+	yInv = fp.BatchInvert(yInv)
+	for k := 0; k < n; k++ {
 		xNegOverY[k].Mul(&P[k].X, &yInv[k]).
 			Neg(&xNegOverY[k])
 	}