# Description

This PR refactors KZG to use fixed-argument pairings. This requires to change the SRS verifying key to store the precomputed lines instead of the points G2 and [a]G2. 

TODO: 
- [x] BN254
- [x] BLS12-381
- [x] BLS12-377
- [x] BW6-761
- [x] BLS12-378
- [x] BW6-756
- [x] BLS24-317
- [x] BLS12-315
- [x] BW6-633

~I need to implement an ate variant Miller loop for BW6-633 to implement this feature.~ Done in [07b87ae](https://github.com/Consensys/gnark-crypto/pull/466/commits/07b87aea9e25c7ca3e62d417b7df8fd4fc9f22ae).

## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to not work as expected)

# How has this been tested?

Same tests for KZG work here as well + testing that `Pair` results with fixed and variable arguments match.

# How has this been benchmarked?

Some benchmarks on a AWS `z1d.large` machine:
```
goos: linux
goarch: amd64
pkg: github.com/consensys/gnark-crypto/ecc/*/kzg
cpu: Intel(R) Xeon(R) Platinum 8151 CPU @ 3.40GHz
```
- BN254
```
benchmark                          old ns/op      new ns/op      delta
BenchmarkSRSGen/real_SRS-2         3101757295     3085342398     -0.53%
BenchmarkSRSGen/quick_SRS-2        952900         1523358        +59.87%
BenchmarkKZGCommit/real_SRS-2      122914358      121965943      -0.77%
BenchmarkKZGCommit/quick_SRS-2     114308031      113640429      -0.58%
BenchmarkDivideByXMinusA-2         103447328      102627372      -0.79%
BenchmarkKZGOpen-2                 124670902      124335198      -0.27%
BenchmarkKZGVerify-2               658898         537151         -18.48%
BenchmarkKZGBatchOpen10-2          136281734      136385577      +0.08%
BenchmarkKZGBatchVerify10-2        1043165        922205         -11.60%
BenchmarkToLagrangeG1-2            5548069025     4977902867     -10.28%
```

- BW6-761
```
benchmark                          old ns/op       new ns/op       delta
BenchmarkSRSGen/real_SRS-2         26341215560     26358001933     +0.06%
BenchmarkSRSGen/quick_SRS-2        3537481         10778989        +204.71%
BenchmarkKZGCommit/real_SRS-2      1186236493      1189108978      +0.24%
BenchmarkKZGCommit/quick_SRS-2     1099211577      1102560337      +0.30%
BenchmarkDivideByXMinusA-2         167096682       167218972       +0.07%
BenchmarkKZGOpen-2                 1186146836      1195299942      +0.77%
BenchmarkKZGVerify-2               3779093         3044281         -19.44%
BenchmarkKZGBatchOpen10-2          1200625617      1202907566      +0.19%
BenchmarkKZGBatchVerify10-2        6456328         5730700         -11.24%
BenchmarkToLagrangeG1-2            44017779311     44044702422     +0.06%
```

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I did not modify files generated from templates
- [x] `golangci-lint` does not output errors locally
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules